# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "\u8ba2\u5355\u7cfb\u7edf",
        "version": "v0.0.1"
    },
    "basePath": "/lsx/public/index.php/v1",
    "paths": {
...
        "/lsx/public/index.php/v1/user/add": {
            "post": {
                "tags": [
                    "\u7528\u6237\u6a21\u5757"
                ],
                "summary": "\u589e\u52a0/\u66f4\u65b0\u5458\u5de5",
                "description": "\u589e\u52a0/\u66f4\u65b0\u5458\u5de5\uff0c\u6709employeeid\u4e3a\u66f4\u65b0\uff0c\u6ca1\u6709\u4e3a\u65b0\u589e",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "\u5458\u5de5\u59d3\u540d",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "age",
                        "in": "formData",
                        "description": "\u5458\u5de5\u5e74\u9f84",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "mobile",
                        "in": "formData",
                        "description": "\u624b\u673a\u53f7",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "salary",
                        "in": "formData",
                        "description": "\u85aa\u8d44",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "pwd",
                        "in": "formData",
                        "description": "\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "avatar",
                        "in": "formData",
                        "description": "\u5934\u50cfurl",
                        "type": "string"
                    },
                    {
                        "name": "employeeid",
                        "in": "formData",
                        "description": "\u7528\u6237id\uff0c\u82e5\u4f20\u4e86\u8fd9\u4e2a\u53c2\u6570\uff0c\u5219\u4e3a\u66f4\u65b0\uff0c\u4e0d\u4f20\u5219\u4e3a\u65b0\u589e\u7528\u6237",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
        "/lsx/public/index.php/v1/user/del": {
            "post": {
                "tags": [
                    "\u7528\u6237\u6a21\u5757"
                ],
                "summary": "\u5220\u9664\u7528\u6237\u63a5\u53e3",
                "description": "\u901a\u8fc7\u7528\u6237id\u5220\u9664\u7528\u6237\u63a5\u53e3",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "employeeId",
                        "in": "formData",
                        "description": "\u7528\u6237id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
        "/resetPassword": {
            "post": {
                "tags": [
                    "\u7528\u6237\u6a21\u5757"
                ],
                "summary": "\u4fee\u6539\u5bc6\u7801",
                "description": "\u4fee\u6539\u5bc6\u7801",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "newpassword",
                        "in": "formData",
                        "description": "\u65b0\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "\u7528\u6237token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "\u7528\u6237\u6a21\u5757"
                ],
                "summary": "\u767b\u9646\u63a5\u53e3",
                "description": "\u53d1\u9001\u96c7\u5458id\u548c\u5bc6\u7801\uff0c\u8fd4\u56detoken",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "employeeid",
                        "in": "formData",
                        "description": "\u96c7\u5458id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "pwd",
                        "in": "formData",
                        "description": "\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
        "/user?token={token}": {
            "get": {
                "tags": [
                    "\u7528\u6237\u6a21\u5757"
                ],
                "summary": "\u83b7\u53d6\u4e2a\u4eba\u4fe1\u606f\u63a5\u53e3",
                "description": "\u83b7\u53d6\u4e2a\u4eba\u4fe1\u606f\u63a5\u53e3\uff0c\u5982\u679ctoken\u5931\u6548\u5219\u4f1a\u8fd4\u56deHTTP CODE 500",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
...
        "/img/upload": {
            "get": {
                "tags": [
                    "\u7528\u6237\u6a21\u5757"
                ],
                "summary": "\u4e0a\u4f20\u7528\u6237\u5934\u50cf\u8ddf\u5728url\u540e\u9762\u7684\u63cf\u8ff0",
                "description": "\u4e0a\u4f20\u7528\u6237\u5934\u50cf",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "img",
                        "in": "formData",
                        "description": "\u4e0a\u4f20\u7684\u6587\u4ef6",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        }
    },
...
}